backend/requirements.txt
Django>=5.2.5
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
drf-yasg>=1.21.10
channels>=4.0.0
psycopg2-binary>=2.9.0
Pillow>=11.0.0
python-decouple>=3.8
channels-redis>=4.2.0
daphne>=4.1.0
celery>=5.3.0
django-celery-beat>=2.5.0
redis>=5.0.0
django-redis>=5.4.0
async-timeout>=5.0.1
flower>=2.0.0